refactor: useModal에서 모달 렌더링 기능 추가
